+++ data_structures/external_memory_node.cpp
@@ -30,7 +30,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 #include <limits>
 
 ExternalMemoryNode::ExternalMemoryNode(
-    int lat, int lon, int ele, unsigned int node_id, bool barrier, bool traffic_lights)
+    int lat, int lon, double ele, unsigned int node_id, bool barrier, bool traffic_lights)
     : QueryNode(lat, lon, ele, node_id), barrier(barrier), traffic_lights(traffic_lights)
 {
 }
@@ -39,14 +39,14 @@ ExternalMemoryNode::ExternalMemoryNode() : barrier(false), traffic_lights(false)
 
 ExternalMemoryNode ExternalMemoryNode::min_value()
 {
-    return ExternalMemoryNode(0, 0, 0, 0, false, false);
+    return ExternalMemoryNode(0, 0, 0.0, 0, false, false);
 }
 
 ExternalMemoryNode ExternalMemoryNode::max_value()
 {
     return ExternalMemoryNode(std::numeric_limits<int>::max(),
                               std::numeric_limits<int>::max(),
-                              std::numeric_limits<int>::max(),
+                              std::numeric_limits<double>::max(),
                               std::numeric_limits<unsigned>::max(),
                               false,
                               false);
